+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.0.2 - 04-15-2018
+### ADDED
+* Search for Artists
+* Search for Tracks
+* Search for Albums
+* Search for Playlists
+* Search for Artists + Tracks
+* Search for Artists + Albums
+
 # 0.0.1 - 04-15-2018
 ### ADDED
 - User and Public API authentication